refactor: centralize rich stub
