+++ tests/conftest.py
@@ -2,6 +2,13 @@ import os
 import sys
 import types
 
+try:
+    import rich  # noqa: F401
+except Exception:
+    from .rich_stub import register_rich_stub
+
+    register_rich_stub()
+
 # Ensure project root is on the import path for tests
 ROOT = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
 if ROOT not in sys.path:
